Updated files for following along
I decided to go ahead and put the files into their respective folders since I am using brackets to live preview them. for now this will do and is easier to keep track of. Unsure how this will translate to when I am in the field but we will cross that bridge when we get there.